+++ bytes/bytes/Extensions/Array+Shuffled.swift
@@ -1,0 +1,38 @@
+//
+//  Array+shuffle.swift
+//  bytes
+//
+//  Created by Cornelius Horstmann on 13.05.17.
+//  Copyright © 2017 TBO INTERACTIVE GmbH & Co. KG. All rights reserved.
+//
+
+import Foundation
+
+extension NSArray {
+    /// Returns a new array with all the value shuffled.
+    /// It uses the Fisher-Yates algorithm.
+    ///
+    /// [source](https://gist.github.com/natecook1000/0ac03efe07f647b46dae)
+    func shuffled() -> NSArray {
+        let array = self as Array
+        return array.shuffled() as NSArray
+    }
+}
+
+extension MutableCollection where Self.IndexDistance == Int, Self.Index == Int {
+    /// Returns a new array with all the value shuffled.
+    /// It uses the Fisher-Yates algorithm.
+    ///
+    /// [source](https://gist.github.com/natecook1000/0ac03efe07f647b46dae)
+    func shuffled() -> Self {
+        var copy = self
+        let count = self.count
+        for sourceIndex in 0..<(count - 1) {
+            let destinationIndex = Int(arc4random_uniform(UInt32(count - sourceIndex))) + sourceIndex
+            if sourceIndex != destinationIndex {
+                swap(&copy[sourceIndex], &copy[destinationIndex])
+            }
+        }
+        return copy
+    }
+}