+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 * **bugfix** Exposed UIColor hex convenience initializer
 
+* **feature** Added function to shuffle an Array/NSArray.
+
 ## 0.2
 * **feature** Added AsyncViewController
 * **feature** Added RemoteSettings to fetch and store remote configurations